try fix tenant delete bug

+++ modules/map-tenancy-management/src/SharpAbp.Abp.MapTenancyManagement.Application/SharpAbp/Abp/MapTenancyManagement/HybridMapTenantAppService.cs
@@ -238,7 +238,15 @@ namespace SharpAbp.Abp.MapTenancyManagement
         public virtual async Task DeleteAsync(Guid id)
         {
             var mapTenant = await MapTenantRepository.GetAsync(id);
+            if (mapTenant == null)
+            {
+                return;
+            }
             var tenant = await TenantRepository.GetAsync(mapTenant.TenantId);
+            if (tenant == null)
+            {
+                return;
+            }
             await TenantRepository.HardDeleteAsync(tenant);
             await MapTenantRepository.DeleteAsync(mapTenant);
         }
